refactoring container stuff to a wrapper
