+++ src/frontend/src/pages/Auth/ChangePassword.tsx
@@ -1,8 +1,6 @@
 import { Trans, t } from '@lingui/macro';
 import {
   Button,
-  Center,
-  Container,
   Divider,
   Group,
   Paper,
@@ -16,12 +14,11 @@ import { useNavigate } from 'react-router-dom';
 
 import { api } from '../../App';
 import { StylishText } from '../../components/items/StylishText';
-import { ProtectedRoute } from '../../components/nav/Layout';
-import { LanguageContext } from '../../contexts/LanguageContext';
 import { ApiEndpoints } from '../../enums/ApiEndpoints';
 import { clearCsrfCookie } from '../../functions/auth';
 import { apiUrl } from '../../states/ApiState';
 import { useUserState } from '../../states/UserState';
+import { Wrapper } from './LoginLayoutComponent';
 
 export default function Set_Password() {
   const simpleForm = useForm({
@@ -102,52 +99,44 @@ export default function Set_Password() {
   }
 
   return (
-    <LanguageContext>
-      <ProtectedRoute>
-        <Center mih='100vh'>
-          <Container w='md' miw={425}>
-            <Stack>
-              <StylishText size='xl'>{t`Reset Password`}</StylishText>
-              <Divider />
-              {user.username() && (
-                <Paper>
-                  <Group>
-                    <StylishText size='md'>{t`User`}</StylishText>
-                    <Text>{user.username()}</Text>
-                  </Group>
-                </Paper>
-              )}
-              <Divider />
-              <Stack gap='xs'>
-                <PasswordInput
-                  required
-                  aria-label='password'
-                  label={t`Current Password`}
-                  description={t`Enter your current password`}
-                  {...simpleForm.getInputProps('current_password')}
-                />
-                <PasswordInput
-                  required
-                  aria-label='input-password-1'
-                  label={t`New Password`}
-                  description={t`Enter your new password`}
-                  {...simpleForm.getInputProps('new_password1')}
-                />
-                <PasswordInput
-                  required
-                  aria-label='input-password-2'
-                  label={t`Confirm New Password`}
-                  description={t`Confirm your new password`}
-                  {...simpleForm.getInputProps('new_password2')}
-                />
-              </Stack>
-              <Button type='submit' onClick={handleSet}>
-                <Trans>Confirm</Trans>
-              </Button>
-            </Stack>
-          </Container>
-        </Center>
-      </ProtectedRoute>
-    </LanguageContext>
+    <Wrapper>
+      <StylishText size='xl'>{t`Reset Password`}</StylishText>
+      <Divider />
+      {user.username() && (
+        <Paper>
+          <Group>
+            <StylishText size='md'>{t`User`}</StylishText>
+            <Text>{user.username()}</Text>
+          </Group>
+        </Paper>
+      )}
+      <Divider />
+      <Stack gap='xs'>
+        <PasswordInput
+          required
+          aria-label='password'
+          label={t`Current Password`}
+          description={t`Enter your current password`}
+          {...simpleForm.getInputProps('current_password')}
+        />
+        <PasswordInput
+          required
+          aria-label='input-password-1'
+          label={t`New Password`}
+          description={t`Enter your new password`}
+          {...simpleForm.getInputProps('new_password1')}
+        />
+        <PasswordInput
+          required
+          aria-label='input-password-2'
+          label={t`Confirm New Password`}
+          description={t`Confirm your new password`}
+          {...simpleForm.getInputProps('new_password2')}
+        />
+      </Stack>
+      <Button type='submit' onClick={handleSet}>
+        <Trans>Confirm</Trans>
+      </Button>
+    </Wrapper>
   );
 }
