+++ src/frontend/src/pages/Auth/LoggedIn.tsx
@@ -1,10 +1,11 @@
 import { Trans } from '@lingui/macro';
-import { Card, Container, Group, Loader, Stack, Text } from '@mantine/core';
+import { Group, Loader, Text } from '@mantine/core';
 import { useDebouncedCallback } from '@mantine/hooks';
 import { useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 
 import { checkLoginState } from '../../functions/auth';
+import { Wrapper } from './LoginLayoutComponent';
 
 export default function Logged_In() {
   const navigate = useNavigate();
@@ -17,19 +18,13 @@ export default function Logged_In() {
   }, [navigate]);
 
   return (
-    <Container>
-      <Stack align='center'>
-        <Card shadow='sm' padding='lg' radius='md'>
-          <Stack>
-            <Text size='lg'>
-              <Trans>Checking if you are already logged in</Trans>
-            </Text>
-            <Group justify='center'>
-              <Loader />
-            </Group>
-          </Stack>
-        </Card>
-      </Stack>
-    </Container>
+    <Wrapper>
+      <Text size='lg'>
+        <Trans>Checking if you are already logged in</Trans>
+      </Text>
+      <Group justify='center'>
+        <Loader />
+      </Group>
+    </Wrapper>
   );
 }
