+++ src/frontend/src/router.tsx
@@ -7,6 +7,10 @@ import { Loadable } from './functions/loading';
 export const LayoutComponent = Loadable(
   lazy(() => import('./components/nav/Layout'))
 );
+export const LoginLayoutComponent = Loadable(
+  lazy(() => import('./pages/Auth/LoginLayoutComponent'))
+);
+
 export const Home = Loadable(lazy(() => import('./pages/Index/Home')));
 
 export const CompanyDetail = Loadable(
@@ -103,17 +107,18 @@ export const AdminCenter = Loadable(
 export const NotFound = Loadable(
   lazy(() => import('./components/errors/NotFound'))
 );
+
+// Auth
 export const Login = Loadable(lazy(() => import('./pages/Auth/Login')));
-export const MFALogin = Loadable(lazy(() => import('./pages/Auth/MFALogin')));
-export const MFASetup = Loadable(lazy(() => import('./pages/Auth/MFASetup')));
+export const LoggedIn = Loadable(lazy(() => import('./pages/Auth/LoggedIn')));
 export const Logout = Loadable(lazy(() => import('./pages/Auth/Logout')));
-export const Logged_In = Loadable(lazy(() => import('./pages/Auth/Logged-In')));
-export const Reset = Loadable(lazy(() => import('./pages/Auth/Reset')));
+export const Mfa = Loadable(lazy(() => import('./pages/Auth/MFA')));
+export const MfaSetup = Loadable(lazy(() => import('./pages/Auth/MFASetup')));
 
 export const ChangePassword = Loadable(
   lazy(() => import('./pages/Auth/ChangePassword'))
 );
-
+export const Reset = Loadable(lazy(() => import('./pages/Auth/Reset')));
 export const ResetPassword = Loadable(
   lazy(() => import('./pages/Auth/ResetPassword'))
 );
@@ -173,16 +178,20 @@ export const routes = (
         <Route path='customer/:id/*' element={<CustomerDetail />} />
       </Route>
     </Route>
-    <Route path='/' errorElement={<ErrorPage />}>
+    <Route
+      path='/'
+      element={<LoginLayoutComponent />}
+      errorElement={<ErrorPage />}
+    >
       <Route path='/login' element={<Login />} />,
       <Route path='/register' element={<Login />} />,
-      <Route path='/mfa' element={<MFALogin />} />,
-      <Route path='/mfa-setup' element={<MFASetup />} />,
+      <Route path='/logged-in' element={<LoggedIn />} />
       <Route path='/logout' element={<Logout />} />,
-      <Route path='/logged-in' element={<Logged_In />} />
+      <Route path='/mfa' element={<Mfa />} />,
+      <Route path='/mfa-setup' element={<MfaSetup />} />,
+      <Route path='/change-password' element={<ChangePassword />} />
       <Route path='/reset-password' element={<Reset />} />
       <Route path='/set-password' element={<ResetPassword />} />
-      <Route path='/change-password' element={<ChangePassword />} />
       <Route path='/verify-email/:key' element={<VerifyEmail />} />
     </Route>
   </Routes>
